esm: move hooks handling into separate class
