+++ lib/internal/modules/esm/load.js
@@ -5,6 +5,7 @@ const {
   RegExpPrototypeExec,
   decodeURIComponent,
 } = primordials;
+const { kEmptyObject } = require('internal/util');
 
 const { defaultGetFormat } = require('internal/modules/esm/get_format');
 const { validateAssertions } = require('internal/modules/esm/assert');
@@ -22,6 +23,7 @@ const { Buffer: { from: BufferFrom } } = require('buffer');
 const { URL } = require('internal/url');
 const {
   ERR_INVALID_URL,
+  ERR_UNKNOWN_MODULE_FORMAT,
   ERR_UNSUPPORTED_ESM_URL_SCHEME,
 } = require('internal/errors').codes;
 
@@ -69,7 +71,7 @@ async function getSource(url, context) {
  * @param {object} context
  * @returns {object}
  */
-async function defaultLoad(url, context) {
+async function defaultLoad(url, context = kEmptyObject) {
   let responseURL = url;
   const { importAssertions } = context;
   let {
@@ -100,6 +102,27 @@ async function defaultLoad(url, context) {
   };
 }
 
+
+/**
+ * For a falsy `format` returned from `load`, throw an error.
+ * This could happen from either a custom user loader _or_ from the default loader, because the default loader tries to
+ * determine formats for data URLs.
+ * @param {string} url The resolved URL of the module
+ * @param {null | undefined | false | 0 | -0 | 0n | ''} format Falsy format returned from `load`
+ */
+function throwUnknownModuleFormat(url, format) {
+  const dataUrl = RegExpPrototypeExec(
+    /^data:([^/]+\/[^;,]+)(?:[^,]*?)(;base64)?,/,
+    url,
+  );
+
+  throw new ERR_UNKNOWN_MODULE_FORMAT(
+    dataUrl ? dataUrl[1] : format,
+    url);
+}
+
+
 module.exports = {
   defaultLoad,
+  throwUnknownModuleFormat,
 };
