+++ lib/internal/process/esm_loader.js
@@ -16,7 +16,7 @@ const esmLoader = new ESMLoader();
 exports.esmLoader = esmLoader;
 
 // Module.runMain() causes loadESM() to re-run (which it should do); however, this should NOT cause
-// ESM to be re-initialised; doing so causes duplicate custom loaders to be added to the public
+// ESM to be re-initialized; doing so causes duplicate custom loaders to be added to the public
 // esmLoader.
 let isESMInitialized = false;
 
@@ -80,7 +80,7 @@ function loadModulesInIsolation(parentURL, specifiers, loaders = []) {
   internalEsmLoader.addCustomLoaders(loaders);
   internalEsmLoader.preload();
 
-  // Importation must be handled by internal loader to avoid poluting userland
+  // Importation must be handled by internal loader to avoid polluting userland
   return internalEsmLoader.import(
     specifiers,
     parentURL,
